requirements.simple.txt
Flask==2.2.5
Flask-SQLAlchemy==2.5.1
Flask-Login==0.6.2
Flask-WTF==1.1.1
Flask-Mail==0.9.1
Flask-Migrate==4.0.4
Flask-Session==0.5.0
Flask-Talisman==1.0.0
PyMySQL==1.0.2
gunicorn==20.1.0
Werkzeug==2.2.3
SQLAlchemy==1.4.23
WTForms==3.0.1
email-validator==1.1.3
python-dotenv==0.19.0
cryptography==3.4.7
PyPDF2==1.26.0
python-barcode==0.15.1
reportlab==4.0.4
Pillow==10.0.0
pywin32==306; sys_platform == 'win32' 